Moved function to sherlock-utils, reverting utils folder
