+++ extensions/sherlock-utils/index.ts
@@ -1,4 +1,5 @@
 export * from './src/derivable-cache';
+export * from './src/function-data-source';
 export * from './src/from-promise';
 export * from './src/lift';
 export * from './src/pairwise';
