+++ src/derivable/base-derivable.ts
@@ -1,9 +1,7 @@
 import { Derivable, MaybeFinalState, SettableDerivable } from '../interfaces';
 import { autoCacheMode, connect, disconnect, finalize, internalGetState, observers } from '../symbols';
 import { independentTracking, isRecordingObservations, maybeDisconnectInNextTick, TrackedObservable, TrackedObserver } from '../tracking';
-import { FinalWrapper } from '../utils/final-wrapper';
-import { prepareCreationStack } from '../utils/augment-stack';
-import { uniqueId } from '../utils/unique-id';
+import { FinalWrapper, prepareCreationStack, uniqueId } from '../utils';
 
 /**
  * The base class for all Derivables. Derivables must extend from this, to be 'tracked' and to classify as a Derivable.
